refactor 중복 예약 신청 시 예약 status_id = 3으로 변환 구현

+++ linkup/src/main/java/com/learningcrew/linkup/place/command/domain/aggregate/entity/PlaceReview.java
@@ -24,7 +24,7 @@ public class PlaceReview {
     private Integer placeId;
 
     @Column(name = "participation_id", nullable = false)
-    private Integer participationId;
+    private Long participationId;
 
     @Column(name = "status_id", nullable = false)
     private Integer statusId;
